refactor monolithic post full details component
1. Modify post full details to take optional post params and proposalType and index
2. If the post prop is provided do a initial load
3. Fetch proposal and do a sync when proposal is fetched
