+++ lib/components/feed/index.ts
@@ -1,6 +1,5 @@
 import CommentBox from "./CommentBox";
 import CommentCard from "./CommentCard";
 import CommentList from "./CommentList";
-import PostFullDetails from "./postFullDetails";
 
-export { CommentBox, CommentCard, CommentList, PostFullDetails } 
+export { CommentBox, CommentCard, CommentList }
