+++ lib/components/proposal/full-details/index.ts
@@ -1,0 +1,3 @@
+import PostFullDetails from "./PostFullDetails";
+
+export default PostFullDetails
